--- back/src/main/java/com/openclassrooms/mddapi/configuration/WebConfig.java
@@ -1,16 +1,0 @@
-package com.openclassrooms.mddapi.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-// @Configuration
-// public class WebConfig implements WebMvcConfigurer {
-    
-//     // TODO : voir avec SpringSecurity 
-//     // --> https://www.baeldung.com/spring-cors
-//     // @Override
-//     // public void addCorsMappings(CorsRegistry registry) {
-//     //     registry.addMapping("http://localhost:4200");
-//     // }
-// }
