+++ back/src/main/java/com/openclassrooms/mddapi/controller/UserController.java
@@ -15,40 +15,40 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.openclassrooms.mddapi.dto.UserDto;
 import com.openclassrooms.mddapi.exception.UserNotFoundException;
-import com.openclassrooms.mddapi.service.IUserService;
+import com.openclassrooms.mddapi.service.UserService;
 
 @RestController
 @RequestMapping("/api/user")
 public class UserController {
 
-    private IUserService userService;
+    private UserService userService;
 
     private static final Logger log = LoggerFactory.getLogger(UserController.class);
 
-    public UserController(IUserService userService) {
+    public UserController(UserService userService) {
         this.userService = userService;
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<UserDto> findById(@PathVariable("id") String id) throws NumberFormatException, UserNotFoundException {
         
+        log.info("(get) /user/{} : Searching user with id {}", id, id);
+
         UserDto userDto = this.userService.findById(Long.valueOf(id));
         
         return ResponseEntity.status(HttpStatus.OK).body(userDto);
-
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<UserDto> update(@PathVariable("id") String id, 
     @Valid @ModelAttribute("UserDto") UserDto userDto) throws UserNotFoundException {
 
-        log.info("/user/{id} : Trying to update user with id {}", id);
+        log.info("(put) /user/{} : Trying to update user with id {}", id, id);
 
         userDto.setId(Long.valueOf(id));
         UserDto updatedUserDto = userService.update(userDto);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(updatedUserDto);
-
     }
     
 }
