+++ back/src/main/java/com/openclassrooms/mddapi/dto/CommentDto.java
@@ -76,6 +76,5 @@ public class CommentDto {
     public void setUpdatedAt(LocalDate updatedAt) {
         this.updatedAt = updatedAt;
     }
-
-    
+ 
 }
