+++ back/src/main/java/com/openclassrooms/mddapi/dto/LoginDto.java
@@ -1,6 +1,6 @@
 package com.openclassrooms.mddapi.dto;
 
-public class LoginRegisterDto {
+public class LoginDto {
 
         private String email;
