+++ back/src/main/java/com/openclassrooms/mddapi/dto/PostRegisterDto.java
@@ -3,13 +3,9 @@ package com.openclassrooms.mddapi.dto;
 public class PostRegisterDto {
     
     private String topic;
-
     private String title;
-
     private String content;
-
     private UserDto auteur;
-
 
     public String getTopic() {
         return this.topic;
@@ -54,6 +50,5 @@ public class PostRegisterDto {
             ", auteur='" + getAuteur() + "'" +
             "}";
     }
-
 
 }
