+++ back/src/main/java/com/openclassrooms/mddapi/model/Comment.java
@@ -31,7 +31,7 @@ public class Comment {
 	@JoinColumn(name = "post_id")
 	private Post post;
 
-	@Column
+	@Column(columnDefinition="TEXT")
 	@NotBlank
 	private String content;
