--- back/src/main/java/com/openclassrooms/mddapi/service/ICommentsService.java
@@ -1,17 +1,0 @@
-package com.openclassrooms.mddapi.service;
-
-import java.util.List;
-
-import com.openclassrooms.mddapi.dto.CommentDto;
-import com.openclassrooms.mddapi.dto.PostDto;
-import com.openclassrooms.mddapi.dto.UserDto;
-import com.openclassrooms.mddapi.model.Post;
-import com.openclassrooms.mddapi.model.User;
-
-public interface ICommentsService {
-    
-    public List<CommentDto> getAll(Post post);
-
-    public CommentDto create(String content, UserDto auteurd, PostDto post);
-    
-}
