--- back/src/main/java/com/openclassrooms/mddapi/service/ITopicService.java
@@ -1,18 +1,0 @@
-package com.openclassrooms.mddapi.service;
-
-import java.util.List;
-
-import com.openclassrooms.mddapi.dto.TopicDto;
-
-public interface ITopicService {
-
-	/**
-	 * Returns a list of all topics.
-	 * 
-	 * @return List<TopicDto>
-	 */
-	List<TopicDto> getAllTopics();
-
-	public TopicDto getTopicById(Long id);
-
-}
